Create custom hook success
